src/requirements.txt
PyGithub==1.55
boto3==1.18.17